task_3/requirements.txt
accelerate==0.31.0
torch==2.3.1
Pillow==9.5.0
transformers==4.41.2
diffusers==0.29.0